add change template to spa

+++ app/Http/Controllers/CompanyController.php
@@ -239,6 +239,8 @@ class CompanyController extends Controller {
         $input['banner'] = $img_banner;
         $input['images'] = $allPic;
         $input['user'] = \Auth::user()->id;
+        $input['email'] = \Auth::user()->email;
+        $input['phone'] = \Auth::user()->phone;
         
         $company = Company::create($input);
 
@@ -460,6 +462,8 @@ class CompanyController extends Controller {
                         'companies.sologan', 
                         'companies.description',
                         'companies.images',
+                        'companies.template',
+                        'companies.site_url',
                         'users.phone as hotline'
                 )
                 ->where('companies.id', $id)
@@ -480,9 +484,132 @@ class CompanyController extends Controller {
 
             $star = intval($totalStar / $numberComment);
 
-            return view('company.info', array('company' => $company, 'company_id' => $company_id, 'cv_id' => $cv_id, 'followed' => $followed, 'comments' => $comments, 'votes' => $star));
+            $jobs = \DB::table('jobs')
+                    ->join('companies', 'companies.id', '=', 'jobs.company')
+                    ->join('salaries', 'salaries.id', '=', 'jobs.salary')
+                    ->join('cities', 'cities.id', '=', 'companies.city')
+                    ->join('districts', 'districts.id', '=', 'companies.district')
+                    ->where('companies.id', '=', $company->id)
+                    ->select('jobs.id as id', 'jobs.name as name', 'jobs.number as number', 'jobs.expiration_date as expiration_date', 'salaries.name as salary', 'companies.logo', 'companies.name as companyname', 'cities.name as city', 'districts.name as district')
+                    ->orderBy('jobs.created_at', 'desc')
+                    ->take(12)
+                    ->get();
+
+            if($company->template == 0){
+                return view('company.info', array('company' => $company, 'company_id' => $company_id, 'cv_id' => $cv_id, 'followed' => $followed, 'comments' => $comments, 'votes' => $star, 'template' => $company->template, 'jobs' => $jobs));
+            }else if($company->template == 1){
+                return view('company.view01', array('company' => $company, 'company_id' => $company_id, 'cv_id' => $cv_id, 'followed' => $followed, 'comments' => $comments, 'votes' => $star, 'template' => $company->template, 'jobs' => $jobs));
+            }else if($company->template == 2){
+                return view('company.view02', array('company' => $company, 'company_id' => $company_id, 'cv_id' => $cv_id, 'followed' => $followed, 'comments' => $comments, 'votes' => $star, 'template' => $company->template, 'jobs' => $jobs));
+            }else if($company->template == 3){
+                return view('company.view03', array('company' => $company, 'company_id' => $company_id, 'cv_id' => $cv_id, 'followed' => $followed, 'comments' => $comments, 'votes' => $star, 'template' => $company->template, 'jobs' => $jobs));
+            }else{
+                return view('company.info', array('company' => $company, 'company_id' => $company_id, 'cv_id' => $cv_id, 'followed' => $followed, 'comments' => $comments, 'votes' => $star, 'jobs' => $jobs));
+            }
         }
         return view('errors.404');
+    }
+ 
+    public function returnView($id){
+        $company_id = -1;
+        $cv_id = -1;
+         // load company info
+        if (\Auth::check()) {
+            $current_id = \Auth::user()->id;
+            
+            //get company 
+            $my_company = \DB::table('companies')
+                    ->where('companies.user', $current_id)
+                    ->select(
+                        'id'
+                    )
+                    ->first();
+            if($my_company){
+                $company_id = $my_company->id;
+            }
+            
+            //get CV 
+            $cv_user = \DB::table('curriculum_vitaes')
+                    ->where('curriculum_vitaes.user', $current_id)
+                    ->select(
+                        'id'
+                    )
+                    ->first();
+            if($cv_user){
+                $cv_id = $cv_user->id;
+            }
+
+            // check followed
+            $follow = Follow::where('user', $current_id)->where('company', $id)->first();
+            if ($follow)
+                $followed = 1;
+            else
+                $followed = 0;
+        }else {
+            $followed = 0;
+        }
+        // $company = Company::find($id);
+        $company = \DB::table('companies')
+                ->join('cities', 'cities.id', '=', 'companies.city')
+                ->join('districts', 'districts.id', '=', 'companies.district')
+                ->join('towns', 'towns.id', '=', 'companies.town')
+                ->join('company_sizes', 'company_sizes.id', '=', 'companies.size')
+                ->join('users', 'users.id', '=', 'companies.user')
+                ->select(
+                        'companies.id', 
+                        'companies.name', 
+                        'companies.logo', 
+                        'companies.user', 
+                        'companies.banner', 
+                        'companies.youtube_link', 
+                        'companies.lat', 
+                        'companies.lng', 
+                        'companies.address', 
+                        'cities.name as city', 
+                        'districts.name as district', 
+                        'towns.name as town', 
+                        'companies.jobs', 
+                        'company_sizes.size as size', 
+                        'companies.sologan', 
+                        'companies.description',
+                        'companies.images',
+                        'users.phone as hotline'
+                )
+                ->where('companies.id', $id)
+                ->first();
+
+        if ($company) {
+            // load comment of company
+            $comments = Comment::where('company', $id)->get();
+            $totalStar = 0;
+            foreach ($comments as $comment) {
+                $totalStar = $comment->star;
+            }
+
+            if (count($comments) == 0)
+                $numberComment = 1;
+            else
+                $numberComment = count($comments);
+
+            $star = intval($totalStar / $numberComment);
+
+            $jobs = Job::where('company', $id)->paginate(5);
+
+            return array('company' => $company, 'company_id' => $company_id, 'cv_id' => $cv_id, 'followed' => $followed, 'comments' => $comments, 'votes' => $star, 'jobs' => $jobs);
+        }
+        return view('errors.404');
+    }
+
+    public function view01($id) {
+        return view('company.view01', $this->returnView($id));
+    }
+
+    public function view02($id) {
+        return view('company.info', $this->returnView($id));
+    }
+
+    public function view03($id) {
+        return view('company.info', $this->returnView($id));        
     }
 
     public function listjobs($id) {
@@ -556,33 +683,56 @@ class CompanyController extends Controller {
     }
 
     public function follow(Request $request) {
-        $input = $request->all();
-        $current_id = \Auth::user()->id;
+        if (\Auth::check()) {
+            $input = $request->all();
+            $current_id = \Auth::user()->id;
 
-        $follow = Follow::where('user', $current_id)->where('company', $input['company'])->first();
-        if ($follow == null) {
-            // store
-            $follow = new Follow;
-            $follow->user = $current_id;
-            $follow->company = $input['company'];
+            $follow = Follow::where('user', $current_id)->where('company', $input['company'])->first();
+            if ($follow == null) {
+                // store
+                $follow = new Follow;
+                $follow->user = $current_id;
+                $follow->company = $input['company'];
 
-            if ($follow->save()) {
+                if ($follow->save()) {
+                    return \Response::json(array('code' => '200', 'message' => 'success', 'follow' => $follow));
+                }
+            } else {
                 return \Response::json(array('code' => '200', 'message' => 'success', 'follow' => $follow));
             }
-        } else {
-            return \Response::json(array('code' => '200', 'message' => 'success', 'follow' => $follow));
+            return \Response::json(array('code' => '404', 'message' => 'unsuccess'));
+        }else{
+            return \Response::json(array('code' => '401', 'message' => 'unauthen!'));
         }
-        return \Response::json(array('code' => '404', 'message' => 'unsuccess'));
     }
 
     public function unfollow(Request $request) {
+        if (\Auth::check()) {
+            $input = $request->all();
+            $current_id = \Auth::user()->id;
+
+            $follow = Follow::where('user', $current_id)->where('company', $input['company'])->first();
+            if ($follow) {
+                if ($follow->delete()) {
+                    return \Response::json(array('code' => '200', 'message' => 'success', 'follow' => $follow));
+                } else {
+                    return \Response::json(array('code' => '404', 'message' => 'unsuccess'));
+                }
+            }
+        }else{
+            return \Response::json(array('code' => '401', 'message' => 'unauthen!'));
+        }
+    }
+
+    public function changeTemplate(Request $request) {
         $input = $request->all();
         $current_id = \Auth::user()->id;
 
-        $follow = Follow::where('user', $current_id)->where('company', $input['company'])->first();
-        if ($follow) {
-            if ($follow->delete()) {
-                return \Response::json(array('code' => '200', 'message' => 'success', 'follow' => $follow));
+        $company = Company::where('user', $current_id)->first();
+        if ($company && isset($input['template'])) {
+            $company->template = (int)$input['template'];
+            if ($company->save()) {
+                return \Response::json(array('code' => '200', 'message' => 'success'));
             } else {
                 return \Response::json(array('code' => '404', 'message' => 'unsuccess'));
             }
